added interface for parsed JWT
