+++ src/app/utils/utils.ts
@@ -140,36 +140,6 @@ export function compareEnums(
 }
 
 /**
- * Parses and returns the payload of a JWT into a JSON object.
- * For me info see {@link https://jwt.io}.
- * @param token a valid JWT
- */
-export function parseJwt(token): {
-  // keycloak user ID
-  sub: string;
-  // keycloak `exact_username` attribute
-  username: string;
-  // email of keycloak user
-  email: string;
-  // roles according to couchdb format
-  "_couchdb.roles": string[];
-  sid: string;
-  jti: string;
-} {
-  const base64Url = token.split(".")[1];
-  const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
-  const jsonPayload = decodeURIComponent(
-    window
-      .atob(base64)
-      .split("")
-      .map((c) => "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2))
-      .join(""),
-  );
-
-  return JSON.parse(jsonPayload);
-}
-
-/**
  * This is a simple shorthand function to create factories for services.
  * The use case is, when multiple services extend the same class and one of these services will be provided.
  * @param service the token for which a service is provided
